1、修复了特殊情况查询的sql，使其正常工作了。 2、实现了同意和不同意特殊情况的俩种处理。

+++ Manage-end/manage-in/src/main/java/com/fanqie/manage/param/acSure.java
@@ -13,10 +13,10 @@ import lombok.experimental.Accessors;
 public class acSure {
     private String userName;
     private String additionalId;
-    private int isSure;
-    private int isFirst;
-    private int isDoubleLanguage;
-    private int isWeekend;
+    private double isSure;
+    private double isFirst;
+    private double isDoubleLanguage;
+    private double isWeekend;
     private String teachName;
     private String classNames;
 }
